+++ lib/src/runner/browser/browser_manager.dart
@@ -15,6 +15,7 @@ import '../../backend/suite.dart';
 import '../../backend/test_platform.dart';
 import '../../util/multi_channel.dart';
 import '../../util/remote_exception.dart';
+import '../../util/stack_trace_mapper.dart';
 import '../../utils.dart';
 import '../load_exception.dart';
 import 'iframe_test.dart';
@@ -53,7 +54,11 @@ class BrowserManager {
   /// [url] should be an HTML page with a reference to the JS-compiled test
   /// suite. [path] is the path of the original test suite file, which is used
   /// for reporting. [metadata] is the parsed metadata for the test suite.
-  Future<Suite> loadSuite(String path, Uri url, Metadata metadata) {
+  ///
+  /// If [mapper] is passed, it's used to map stack traces for errors coming
+  /// from this test suite.
+  Future<Suite> loadSuite(String path, Uri url, Metadata metadata,
+      {StackTraceMapper mapper}) {
     url = url.replace(fragment: Uri.encodeFull(JSON.encode({
       "metadata": metadata.serialize(),
       "browser": browser.identifier
@@ -95,7 +100,8 @@ class BrowserManager {
       return new Suite(response["tests"].map((test) {
         var testMetadata = new Metadata.deserialize(test['metadata']);
         var testChannel = suiteChannel.virtualChannel(test['channel']);
-        return new IframeTest(test['name'], testMetadata, testChannel);
+        return new IframeTest(test['name'], testMetadata, testChannel,
+            mapper: mapper);
       }), metadata: metadata, path: path);
     });
   }